# api/specs/openapi.yaml
openapi: 3.1.0
info:
  title: 1xVet API
  description: Ветеринарная онлайн-платформа
  version: 0.0.1

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /api/health:
    get:
      tags:
        - health
      operationId: healthCheck
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/auth/register:
    post:
      tags:
        - auth
      operationId: register
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/pets:
    get:
      tags:
        - pet
      operationId: getMyPets
      summary: Get current user's pets
      responses:
        "200":
          description: List of pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PetResponse"

    post:
      tags:
        - pet
      operationId: createPet
      summary: Add a new pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePetRequest"
      responses:
        "201":
          description: Pet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetResponse"

  /api/pets/{petId}:
    get:
      tags:
        - pet
      operationId: getPet
      summary: Get pet by ID
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Pet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetResponse"
        "404":
          description: Pet not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - pet
      operationId: updatePet
      summary: Update pet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePetRequest"
      responses:
        "200":
          description: Pet updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetResponse"
        "404":
          description: Pet not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - pet
      operationId: deletePet
      summary: Delete pet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Pet deleted
        "404":
          description: Pet not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/vet/profile:
    get:
      tags:
        - vet
      operationId: getMyVetProfile
      summary: Get current vet's profile
      responses:
        "200":
          description: Vet profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VetProfileResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - vet
      operationId: updateVetProfile
      summary: Update vet profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVetProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VetProfileResponse"

  /api/vets:
    get:
      tags:
        - vet
      operationId: searchVets
      summary: Search veterinarians
      parameters:
        - name: specialty
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of vets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VetProfileResponse"

  /api/vets/{vetId}:
    get:
      tags:
        - vet
      operationId: getVetProfile
      summary: Get vet profile by ID
      parameters:
        - name: vetId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Vet profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VetProfileResponse"
        "404":
          description: Vet not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/login:
    post:
      tags:
        - auth
      operationId: login
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    # health
    HealthResponse:
      $ref: "health.yaml#/schemas/HealthResponse"

    # auth
    RegisterRequest:
      $ref: "auth.yaml#/schemas/RegisterRequest"
    LoginRequest:
      $ref: "auth.yaml#/schemas/LoginRequest"
    AuthResponse:
      $ref: "auth.yaml#/schemas/AuthResponse"
    UserInfo:
      $ref: "auth.yaml#/schemas/UserInfo"

    # pet
    CreatePetRequest:
      $ref: "pet.yaml#/schemas/CreatePetRequest"
    UpdatePetRequest:
      $ref: "pet.yaml#/schemas/UpdatePetRequest"
    PetResponse:
      $ref: "pet.yaml#/schemas/PetResponse"

    # vet
    VetProfileResponse:
      $ref: "vet.yaml#/schemas/VetProfileResponse"
    UpdateVetProfileRequest:
      $ref: "vet.yaml#/schemas/UpdateVetProfileRequest"

    # common
    ErrorResponse:
      $ref: "common.yaml#/schemas/ErrorResponse"